Bugfix: parsing data object paths gives an incorrect suffix
When a path had multiple suffixes, the wrong suffix was being returned
(i.e. not the last one, ignoring compression suffixes). This was
caused by a bad regex. This replaces using a regex with a pathlib API
call and explict suffix filtering.

Also improved the style of some docstrings in the same module.