Move from setuptools_scm to setuptools-git-versioning
setuptools has recently deprecated legacy versioning, so we moved to
configuring setuptools_scm with the no-guess-dev option. It appears
that a release always adds version bump in addition to the git tag.

e.g. release 1.0.1 reports its version as
1.0.1.post1.dev0+g3859438.d20230308

while 3859438 is the commit that was tagged on master as 1.0.1, the
version has a post-release suffix (and appears to have been
interpreted as a dev release?).

The issue reported here appears to be the same:
https://github.com/pypa/setuptools_scm/issues/812

Trying setuptools-git-versioning instead in the hope that it is less
"clever".