Add rename on copy for collections
Previously the copy function would always copy a source collection
into a destination collection (which must already exist). This is
sometimes not desirable e.g. when we want to copy to a collection with
a new name (which doesn't exist, bur whose parent collection does).

This adds this feature. In doing so, it separates the recursive part
of the copy operation into a new, private function and adds a new
top-level copy function which now handles source and destination path
validation and setup required for the rename case.