Add update-secondary-metadata script.
Add functions to update sample and study metadata on Illumina data
objects in iRODs.

Add analysis platform identification Enums and functions.

Add a class for Illumina components and placeholder classes for
PacBio, ONT and Fluidigm components.

Add support for human chromosome columns to IseqFlowcell and Study ORM
classes.

Move the parse_iso_date function to the cli module for re-use.

Add CLI sub-command support to the locate-data-objects script.

Move consent-withdrawn to a locate-data-objects sub-command.

Add an illumina-updates sub-command to the locate-data-objects script.

Rename _ensure_avus_present to ensure_avus_present as it's now
imported by another module. Change the implementation to remove the
limit that none of the argument AVUs may share an attribute.

Make everything in the scripts subdirectory executable.

Refactor test fixtures to separate MLWH and iRODS setup because the
latter takes a long time and is not always needed for some tests.