Fix incorrect value for sample_id metadata
The value should be taken from the id_sample_lims column of the MLWH
sample table, not the sanger_sample_id column.

This change adds a failing test, expands the tests to include all the
sample metadat that is set and then fixes the failing test by
correcting the error above.