Avoid creating a dirty build environment in CI
Build versions are getting tagged dirty by git describe. The main
culprit is the .dockerignore file which excludes some git-tracked
files from the build. This change removes the exclusions.

An additional cause was the local pyenv installation (both the
installer script and install itself) which are now ignored or removed
prior to the build.