Fix accidental expansion of ACL groups
When updating permissions, any iRODS group was being expanded into its
member iRODS users and those user permissions were also being set.

This is a side-effect of the way iRODS reports permissions and can be
worked around by filtering the ACLs to remove unwanted types of
permission. The filtering API has been added to partisan in 2.8.1,
which is now updated in requirements.