Set a clean PATH and PYTHONPATH
When run with Singularity it's possible to have executables and Python
modules be mounted into the container accidentally e.g. via
Singularity's automatic mounting of home directories.

These changes ensure that the PATH and PYTHONPATH are set explicitly
to reduce the chance of running code other than that included in the
image.