Fix handling of PendingRollbackError when processing large datasets
This exception is raised when the database disconnects mid-operation
and has to be handled outside the ORM.

This change includes some changes to the DB engine init args which
make getting bad/stale connections less likely. It also changes the
main work loop to operate of batches of paths, using a fresh DB
session for each and adds handling for the remaindre of a batch should
an exception occur somewhere in the middle of it. SQLAlchemy errors
are handled specially, by propagating them up to where they can
trigger a rollback and allow work to continue with a new DB session.

This updates the Python version requirement to 3.12 (to get
itertools.batched())